LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-scale-composition-expected.txt


FAIL Compositing: property <transform> underlying [scaleX(2)] from add [scaleX(3)] to add [scaleX(4)] at (-0.5) should be [scaleX(5)] assert_equals: expected "matrix ( 5 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 2.5 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from add [scaleX(3)] to add [scaleX(4)] at (0) should be [scaleX(6)] assert_equals: expected "matrix ( 6 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from add [scaleX(3)] to add [scaleX(4)] at (0.25) should be [scaleX(6.5)] assert_equals: expected "matrix ( 6.5 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3.25 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from add [scaleX(3)] to add [scaleX(4)] at (0.5) should be [scaleX(7)] assert_equals: expected "matrix ( 7 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3.5 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from add [scaleX(3)] to add [scaleX(4)] at (0.75) should be [scaleX(7.5)] assert_equals: expected "matrix ( 7.5 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3.75 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from add [scaleX(3)] to add [scaleX(4)] at (1) should be [scaleX(8)] assert_equals: expected "matrix ( 8 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 4 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from add [scaleX(3)] to add [scaleX(4)] at (1.5) should be [scaleX(9)] assert_equals: expected "matrix ( 9 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 4.5 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from add [scaleY(3)] to add [scaleY(4)] at (-0.5) should be [scaleY(5)] assert_equals: expected "matrix ( 1 , 0 , 0 , 5 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 2.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from add [scaleY(3)] to add [scaleY(4)] at (0) should be [scaleY(6)] assert_equals: expected "matrix ( 1 , 0 , 0 , 6 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from add [scaleY(3)] to add [scaleY(4)] at (0.25) should be [scaleY(6.5)] assert_equals: expected "matrix ( 1 , 0 , 0 , 6.5 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3.25 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from add [scaleY(3)] to add [scaleY(4)] at (0.5) should be [scaleY(7)] assert_equals: expected "matrix ( 1 , 0 , 0 , 7 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from add [scaleY(3)] to add [scaleY(4)] at (0.75) should be [scaleY(7.5)] assert_equals: expected "matrix ( 1 , 0 , 0 , 7.5 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3.75 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from add [scaleY(3)] to add [scaleY(4)] at (1) should be [scaleY(8)] assert_equals: expected "matrix ( 1 , 0 , 0 , 8 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 4 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from add [scaleY(3)] to add [scaleY(4)] at (1.5) should be [scaleY(9)] assert_equals: expected "matrix ( 1 , 0 , 0 , 9 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 4.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from add [scaleZ(3)] to add [scaleZ(4)] at (-0.5) should be [scaleZ(5)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 2.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from add [scaleZ(3)] to add [scaleZ(4)] at (0) should be [scaleZ(6)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 6 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from add [scaleZ(3)] to add [scaleZ(4)] at (0.25) should be [scaleZ(6.5)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 6.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.25 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from add [scaleZ(3)] to add [scaleZ(4)] at (0.5) should be [scaleZ(7)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 7 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from add [scaleZ(3)] to add [scaleZ(4)] at (0.75) should be [scaleZ(7.5)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 7.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.75 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from add [scaleZ(3)] to add [scaleZ(4)] at (1) should be [scaleZ(8)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 8 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from add [scaleZ(3)] to add [scaleZ(4)] at (1.5) should be [scaleZ(9)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 9 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from accumulate [scaleX(3)] to accumulate [scaleX(4)] at (-0.5) should be [scaleX(3.5)] assert_equals: expected "matrix ( 3.5 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 2.5 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from accumulate [scaleX(3)] to accumulate [scaleX(4)] at (0) should be [scaleX(4)] assert_equals: expected "matrix ( 4 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from accumulate [scaleX(3)] to accumulate [scaleX(4)] at (0.25) should be [scaleX(4.25)] assert_equals: expected "matrix ( 4.25 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3.25 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from accumulate [scaleX(3)] to accumulate [scaleX(4)] at (0.5) should be [scaleX(4.5)] assert_equals: expected "matrix ( 4.5 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3.5 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from accumulate [scaleX(3)] to accumulate [scaleX(4)] at (0.75) should be [scaleX(4.75)] assert_equals: expected "matrix ( 4.75 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 3.75 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from accumulate [scaleX(3)] to accumulate [scaleX(4)] at (1) should be [scaleX(5)] assert_equals: expected "matrix ( 5 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 4 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleX(2)] from accumulate [scaleX(3)] to accumulate [scaleX(4)] at (1.5) should be [scaleX(5.5)] assert_equals: expected "matrix ( 5.5 , 0 , 0 , 1 , 0 , 0 ) " but got "matrix ( 4.5 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from accumulate [scaleY(3)] to accumulate [scaleY(4)] at (-0.5) should be [scaleY(3.5)] assert_equals: expected "matrix ( 1 , 0 , 0 , 3.5 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 2.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from accumulate [scaleY(3)] to accumulate [scaleY(4)] at (0) should be [scaleY(4)] assert_equals: expected "matrix ( 1 , 0 , 0 , 4 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from accumulate [scaleY(3)] to accumulate [scaleY(4)] at (0.25) should be [scaleY(4.25)] assert_equals: expected "matrix ( 1 , 0 , 0 , 4.25 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3.25 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from accumulate [scaleY(3)] to accumulate [scaleY(4)] at (0.5) should be [scaleY(4.5)] assert_equals: expected "matrix ( 1 , 0 , 0 , 4.5 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from accumulate [scaleY(3)] to accumulate [scaleY(4)] at (0.75) should be [scaleY(4.75)] assert_equals: expected "matrix ( 1 , 0 , 0 , 4.75 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 3.75 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from accumulate [scaleY(3)] to accumulate [scaleY(4)] at (1) should be [scaleY(5)] assert_equals: expected "matrix ( 1 , 0 , 0 , 5 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 4 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleY(2)] from accumulate [scaleY(3)] to accumulate [scaleY(4)] at (1.5) should be [scaleY(5.5)] assert_equals: expected "matrix ( 1 , 0 , 0 , 5.5 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0 , 4.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (-0.5) should be [scaleZ(3.5)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 2.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0) should be [scaleZ(4)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0.25) should be [scaleZ(4.25)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4.25 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.25 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0.5) should be [scaleZ(4.5)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0.75) should be [scaleZ(4.75)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4.75 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.75 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (1) should be [scaleZ(5)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scaleZ(2)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (1.5) should be [scaleZ(5.5)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 5.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scale(2, 4)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (-0.5) should be [scale3d(2, 4, 2.5)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 2.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 2.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scale(2, 4)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0) should be [scale3d(2, 4, 3)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scale(2, 4)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0.25) should be [scale3d(2, 4, 3.25)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 3.25 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.25 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scale(2, 4)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0.5) should be [scale3d(2, 4, 3.5)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 3.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scale(2, 4)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (0.75) should be [scale3d(2, 4, 3.75)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 3.75 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 3.75 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scale(2, 4)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (1) should be [scale3d(2, 4, 4)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [scale(2, 4)] from accumulate [scaleZ(3)] to accumulate [scaleZ(4)] at (1.5) should be [scale3d(2, 4, 4.5)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 4 , 0 , 0 , 0 , 0 , 4.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 4.5 , 0 , 0 , 0 , 0 , 1 ) "
